backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4
requests==2.31.0
flask-swagger-ui==4.11.1
flasgger==0.9.7.1
PyJWT==2.8.0
Werkzeug==3.0.1
marshmallow==3.20.1
